Unify attributes retrieval for JSON and HTML rendering
